First persistent version with backend and groups integration
Mobile version currently not usable while Android WRT does not allow
file attaching.

The application now it's persistent, it sends the media to a backend, as
well as the notes.  At the beggining it retrieves all the notes from the
backend.

Every note is saved in a group, this group is designed by the Webinos
PZH, so every device connected to a determined PZH will have the same
notes.

When some device has sent or updated a note, the rest of the connected
devices to the PZH will be notified.